requirements.txt
Flask==1.1.1
fuzzywuzzy==0.17.0
Flask_Cors==3.0.8
Flask_Bootstrap==3.3.7.1
Flask_WTF==0.14.3
jsonpickle==1.2
mutagen==1.42.0
spotipy==2.4.4
SQLAlchemy==1.3.8
WTForms==2.2.1
pyaes==1.6.1
ratelimit==2.2.1
requests==2.22.0
discogs_client==2.2.2
tqdm==4.36.1
colorama==0.4.1
alembic==1.1.0
asyncio==3.4.3
aiohttp==3.6.1
pycryptodome>=3.9.0